refactor: short 타입 필드를 int로 통일 및 불필요한 캐스팅 제거

+++ src/main/java/com/campusnavi/backend/official/crawler/failure/entity/CrawlFailure.java
@@ -41,7 +41,7 @@ public class CrawlFailure {
     private LocalDate publishedAt;
 
     @Column(nullable = false)
-    private short retryCount = 0;
+    private int retryCount = 0;
 
     @Column(length = MAX_ERROR_LENGTH)
     private String lastError;
